backend/requirements.txt
tensorflow==2.16.1
numpy<2.0
pandas==2.2.0
fastapi
uvicorn
python-multipart
Pillow
requests
